feat(perfiles): refactorizacion de pruebas unitarias
